Added Items to the game
Items can now be used. They are currently 3 static items, but we will add functionality for the rest soon.

Duration can be changed by the developer on the item, so if we want to make say firevest+ or something, we can simply just upgrade the duration of it to be longer for the future.

Also, one time consumable items like the defroster work as well.