Object pooling implemented
Object pooling is now implemented into the game. We can make as many different eras and pools of enemies as we want now.

Need to implement it for coins/treasure chests still.

Currently you can't have duplicate names of prefabs in pools (Example: Eagle can't be in prehistoric and feudal japan, needs to be eagle-p- and eagle-fj- or something like that on the prefab name itself.