Added functionality for bolt to be spawned in during gameplay
Bolt spawns at every 3rd wave. Also, the amount that the player obtains is based on the multiplier. If the player misses a bolt, it will have a chance to spawn before 2 more waves pass. Basically the minimum between a bolt spawning if the player hasn't collected one is 2 waves.