Fixed respawn glitches
Respawn glitches fixed:

When dying in either burning, hanging , or frozen state and then respawning; the player will no longer be stuck in the ground and have weird hitboxes.

When respawning, the enemy count has been adjusted so that there will be no enemies that randomly spawn before a bonus or timeswap wave happens.